DyLNet/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.1=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asteroid-filterbanks=0.4.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
audioread=3.0.1=pypi_0
beautifulsoup4=4.12.2=py39haa95532_0
blas=1.0=mkl
blis=0.7.11=pypi_0
bottleneck=1.3.5=py39h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
bs4=4.12.2=py38hd3eb1b0_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.12.12=haa95532_0
catalogue=2.0.10=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py39haa95532_0
cloudpathlib=0.16.0=pypi_0
colorama=0.4.6=py39haa95532_0
colorlog=6.8.0=pypi_0
comm=0.2.1=pyhd8ed1ab_0
confection=0.1.4=pypi_0
contourpy=1.2.0=py39h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
cymem=2.0.8=pypi_0
debugpy=1.8.0=py39h99910a6_1
decorator=5.1.1=pyhd8ed1ab_0
docopt=0.6.2=py39haa95532_1
einops=0.7.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=hd77b12b_0
filelock=3.13.1=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
fr-core-news-sm=3.7.0=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
giflib=5.2.1=h8cc25b3_3
gmpy2=2.1.2=py39h7f96b67_0
greenlet=3.0.3=pypi_0
huggingface-hub=0.20.2=pypi_0
hyperpyyaml=1.2.2=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.6=pypi_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
importlib_resources=6.1.1=py39haa95532_1
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.0=pyha63f2e9_0
ipython=8.18.1=pyh7428d3b_3
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
jiwer=3.0.3=pypi_0
joblib=1.2.0=py39haa95532_0
jpeg=9e=h2bbff1b_1
julius=0.2.7=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.1=py39hcbf5309_0
kiwisolver=1.4.4=py39hd77b12b_0
langcodes=3.3.0=pypi_0
lazy-loader=0.3=pypi_0
lerc=3.0=hd77b12b_0
levenshtein=0.23.0=pypi_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libpng=1.6.39=h8cc25b3_0
librosa=0.10.1=pypi_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.1=hd77b12b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lightning=2.1.3=pypi_0
lightning-utilities=0.10.0=pypi_0
llvmlite=0.41.1=pypi_0
lxml=4.9.3=py39h09808a7_0
lz4-c=1.9.4=h2bbff1b_0
mako=1.3.0=pypi_0
markdown-it-py=2.2.0=py39haa95532_1
markupsafe=2.1.3=pypi_0
matplotlib-base=3.8.0=py39h4ed8f06_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.0=py39haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
msgpack=1.0.7=pypi_0
multidict=6.0.4=pypi_0
munkres=1.1.4=py_0
murmurhash=1.0.10=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
numba=0.58.1=pypi_0
numexpr=2.8.7=py39h2cd9be0_0
numpy=1.26.3=py39h055cbcc_0
numpy-base=1.26.3=py39h65a83cf_0
omegaconf=2.3.0=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.2.0=hcfcfb64_1
optuna=3.5.0=pypi_0
packaging=23.1=py39haa95532_0
pandas=2.1.4=py39h4ed8f06_0
parso=0.8.3=pyhd8ed1ab_0
pescador=2.1.0=py_0
pickleshare=0.7.5=py_1003
pillow=10.0.1=py39h045eedc_0
pip=23.3.2=pypi_0
platformdirs=4.1.0=pyhd8ed1ab_0
pooch=1.8.0=pypi_0
praat-textgrids=1.4.0=pypi_0
praatio=4.4.0=pypi_0
preshed=3.0.9=pypi_0
primepy=1.3=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.2=pypi_0
psutil=5.9.8=py39ha55989b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyannote-audio=3.1.1=pypi_0
pyannote-parser=0.8=pypi_0
pyannote-pipeline=3.0.1=pypi_0
pyannote.core=5.0.0=pyhd8ed1ab_0
pyannote.database=5.0.1=pyhd8ed1ab_0
pyannote.metrics=3.2.1=pyhd8ed1ab_0
pycparser=2.21=pypi_0
pydantic=2.6.0=pypi_0
pydantic-core=2.16.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.15.1=py39haa95532_1
pympi-ling=1.70.2=pypi_0
pyparsing=3.0.9=py39haa95532_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-levenshtein=0.23.0=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch-lightning=2.1.3=pypi_0
pytorch-metric-learning=2.4.1=pypi_0
pytz=2023.3.post1=py39haa95532_0
pywin32=306=py39h99910a6_2
pyyaml=6.0.1=py39h2bbff1b_0
pyzmq=25.1.0=py39hd77b12b_0
rapidfuzz=3.6.1=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rich=13.3.5=py39haa95532_0
ruamel-yaml=0.18.5=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
scikit-learn=1.3.0=py39h4ed8f06_1
scipy=1.11.4=py39h309d312_0
semver=3.0.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=69.0.3=pypi_0
shellingham=1.5.0=py39haa95532_0
simplejson=3.17.6=py39h2bbff1b_0
six=1.16.0=pyhd3eb1b0_1
smart-open=6.4.0=pypi_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
soundfile=0.12.1=pypi_0
soupsieve=2.5=py39haa95532_0
soxr=0.3.7=pypi_0
spacy=3.7.2=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
speechbrain=0.5.16=pypi_0
sqlalchemy=2.0.25=pypi_0
sqlite=3.41.2=h2bbff1b_0
srsly=2.4.8=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py39haa95532_0
tabulate=0.9.0=py39haa95532_0
tbb=2021.8.0=h59b6b97_0
tensorboardx=2.6.2.2=pypi_0
thinc=8.2.2=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
torch=2.1.2=pypi_0
torch-audiomentations=0.11.0=pypi_0
torch-pitch-shift=1.2.4=pypi_0
torchaudio=2.1.2=pypi_0
torchmetrics=1.3.0=pypi_0
tornado=6.3.3=py39ha55989b_1
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
typer=0.9.0=py39haa95532_0
typing-extensions=4.7.1=py39haa95532_0
typing_extensions=4.7.1=py39haa95532_0
tzdata=2023d=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_0
unidecode=1.3.8=pypi_0
urllib3=2.1.0=pypi_0
vc=14.2=h21ff451_1
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wasabi=1.1.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
weasel=0.3.4=pypi_0
werpy=2.1.1=pypi_0
wheel=0.41.2=py39haa95532_0
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
yarl=1.9.4=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.17.0=py39haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
